# Add zip_subtrees for paired iteration over DataTrees