… xarray/core/iterators.py