@shoyer 2024-10-15
I made a pass at re-implementing `map_over_datasets` using `zip_subtree` in https://github.com/shoyer/xarray/pull/2