@shoyer 2024-10-15
> > so the result does not depend on the order in which child nodes are defined.
> 
> 
> I'm not sure I understand this. Surely `zip_subtrees` is zipping nodes according to the order they appear in `.children`, which is the order they are defined?

Let me restate this: zip_subtrees allows for zipping together multiple DataTree objects even if child nodes on different trees are defined in different orders, as long as the sets of each node's children match.