@TomNicholas 2024-10-15
And when you say "match" you mean the set of names of the children on tree A match the set of names of the children on tree B?