@TomNicholas 2024-10-16
> matching is based on relative path from each root.

So I think this will basically create a breaking change relative to how `map_over_subtree` (and hence all arithmetic) used to work in `xarray-contrib/datatree`. In the old model the definition of isomorphic was such that you could actually multiply two trees with the same structure but differently-named nodes, e.g.

```python
dt1 = DataTree.from_dict({'a': ..., 'a/b': ..., 'a/c': ...})
dt2 = DataTree.from_dict({'e': ..., 'e/f': ..., 'e/g': ...})

dt1 * dt2  # would return a tree with names following dt1
```

I was never really sure if that generality was actually necessary though.

Your new definition of `isomorphic` is sort of like having `strict_names=True` always, except that the as long as the set of names match then their names will be used to determine their corresponding nodes in the other tree, regardless of order.

Therefore if we're relaxing this then I think there is no longer any need to think of the data model of datatree as being an _ordered_ set of children, as no mapping behaviour will depend on that order any longer.

I think it's okay to change this behaviour, including getting rid of the `strict_names` kwarg. After all I have never once needed the generality above! I just wanted to really spell out the implications to make sure I and anyone reading is following.

The last sentence of [this section](https://github.com/pydata/xarray/blob/main/doc/user-guide/hierarchical-data.rst#comparing-trees-for-isomorphism) of the docs will need changing, and the change would merit an entry in the migration guide.