@shoyer 2024-10-16
> The last sentence of [this section](https://github.com/pydata/xarray/blob/main/doc/user-guide/hierarchical-data.rst#comparing-trees-for-isomorphism) of the docs will need changing, and the change would merit an entry in the migration guide.

I'll do this in the next PR when I migrate over the `map_over_datasets` stuff.